// src/app/models/daily.ts
export interface Daily {
  apparentTemperatureHigh: number;
  apparentTemperatureHighTime: number;
  apparentTemperatureLow: number;
  apparentTemperatureLowTime: number;
  apparentTemperatureMax: number;
  apparentTemperatureMaxTime: number;
  apparentTemperatureMin: number;
  apparentTemperatureMinTime: number;
  cardinal: string;
  cloudCover: number;
  dewPoint: number;
  humidity: number;
  icon: string;
  moonPhase: number;
  ozone: number;
  precipIntensity: number;
  precipIntensityMax: number;
  precipProbability: number;
  pressure: number;
  summary: string;
  sunriseTime: number;
  sunsetTime: number;
  tempMaxC: number;
  tempMinC: number;
  temperatureHigh: number;
  temperatureHighTime: number;
  temperatureLow: number;
  temperatureLowTime: number;
  temperatureMax: number;
  temperatureMaxTime: number;
  temperatureMin: number;
  temperatureMinTime: number;
  time: number;
  uvIndex: number;
  uvIndexTime: number;
  visibility: number;
  windBearing: number;
  windGust: number;
  windGustTime: number;
  windKph: number;
  windSpeed: number;
}